break: rename `worktop/kv` -> `worktop/cfw.kv`
