+++ packages/worktop/src/cfw.kv.assets.ts
@@ -1,5 +1,5 @@
 import { lookup, mimes } from 'mrmime';
-import type { KV } from 'worktop/kv';
+import type { KV } from 'worktop/cfw.kv';
 
 export { mimes };
