+++ packages/worktop/types/durable.ts
@@ -6,7 +6,7 @@ import * as utils from 'worktop/utils';
 import type { Bindings } from 'worktop/cfw';
 import type { Durable } from 'worktop/cfw.durable';
 import type { WebSocket } from 'worktop/ws';
-import type { KV } from 'worktop/kv';
+import type { KV } from 'worktop/cfw.kv';
 
 interface CustomBindings extends Bindings {
 	SECRET: string;
