+++ packages/worktop/types/router.strict.ts
@@ -6,7 +6,7 @@ import * as sw from 'worktop/sw';
 import type { Context } from 'worktop';
 import type { Bindings } from 'worktop/cfw';
 import type { Durable } from 'worktop/cfw.durable';
-import type { KV } from 'worktop/kv';
+import type { KV } from 'worktop/cfw.kv';
 
 interface Custom extends Context {
 	start?: number;
